Split project into multiple files
Gave all function names a prefix that matches their corresponding
filenames
