+++ .gitignore
@@ -1,3 +1,1 @@
 default/
-
-*.aws
