+++ include/delays.c
@@ -1,0 +1,14 @@
+#include "avr_compiler.h"
+#include "delays.h"
+
+// ===========================================================
+// delay1ms() 1ms delay function
+// ===========================================================
+// this wrapper function calls _delay_ms with a known value of 1
+// if you call _delay_ms(variable) then the floating point library
+// is going to be included and your output file gets much larger
+// source: http://efundies.com/accurate-delays-with-avr-in-c/
+void delay1ms(uint16_t ms) {
+    uint16_t i;
+    for(i=0;i<ms;i++) _delay_ms(1);
+}//end of delay1ms()
