+++ include/delays.h
@@ -1,0 +1,1 @@
+void delay1ms(uint16_t ms);
